Simplify (no need for scoped_session), consolidate redundant fixtures
